requirement.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.2.9
python-slugify==8.0.1
requests==2.31.0
SQLAlchemy==2.0.15
text-unidecode==1.3
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.4
WTForms==3.0.1
